+++ src/main/java/org/opensearch/index/codec/customcodecs/backward_codecs/ZstdNoDict95Codec.java
@@ -6,24 +6,25 @@
  * compatible open source license.
  */
 
-package org.opensearch.index.codec.customcodecs;
+package org.opensearch.index.codec.customcodecs.backward_codecs;
 
 import org.opensearch.common.settings.Setting;
+import org.opensearch.index.codec.CodecAliases;
 import org.opensearch.index.codec.CodecSettings;
 import org.opensearch.index.engine.EngineConfig;
 
+import java.util.Set;
+
 /**
- * ZstdDeprecatedCodec provides ZSTD compressor using the <a href="https://github.com/luben/zstd-jni">zstd-jni</a> library.
- * Added to support backward compatibility for indices created with Lucene95CustomCodec as codec name.
+ * ZstdNoDictCodec provides ZSTD compressor without a dictionary support.
  */
-@Deprecated(since = "2.10")
-public class Zstd95DeprecatedCodec extends Lucene95CustomCodec implements CodecSettings {
+public class ZstdNoDict95Codec extends Lucene95CustomCodec implements CodecSettings, CodecAliases {
 
     /**
-     * Creates a new ZstdDefaultCodec instance with the default compression level.
+     * Creates a new ZstdNoDictCodec instance with the default compression level.
      */
-    public Zstd95DeprecatedCodec() {
-        super(Mode.ZSTD_DEPRECATED);
+    public ZstdNoDict95Codec() {
+        super(Mode.ZSTD_NO_DICT);
     }
 
     /** The name for this codec. */
@@ -35,5 +36,10 @@ public class Zstd95DeprecatedCodec extends Lucene95CustomCodec implements CodecS
     @Override
     public boolean supports(Setting<?> setting) {
         return setting.equals(EngineConfig.INDEX_CODEC_COMPRESSION_LEVEL_SETTING);
+    }
+
+    @Override
+    public Set<String> aliases() {
+        return Mode.ZSTD_NO_DICT.getAliases();
     }
 }
