+++ src/main/resources/META-INF/services/org.apache.lucene.codecs.Codec
@@ -1,5 +1,5 @@
-org.opensearch.index.codec.customcodecs.Zstd95Codec
-org.opensearch.index.codec.customcodecs.ZstdNoDict95Codec
-org.opensearch.index.codec.customcodecs.Zstd95DeprecatedCodec
+org.opensearch.index.codec.customcodecs.backward_codecs.Zstd95Codec
+org.opensearch.index.codec.customcodecs.backward_codecs.ZstdNoDict95Codec
+org.opensearch.index.codec.customcodecs.backward_codecs.Zstd95DeprecatedCodec
 org.opensearch.index.codec.customcodecs.Zstd99Codec
 org.opensearch.index.codec.customcodecs.ZstdNoDict99Codec
